Change the doorlock over from serial to GPIO.
Updating to use a GPIO library rather than a serial connection to the door as we've nixed the necessity of using the Arduino as a middleman. No change should need to be made in the chimera/reader library for this to work other than recompiling.

This will only work with linux/arm, so make sure to build as follows:

    GOOS=linux GOARCH=arm go build